requirements.txt
# Data collection
pytrends>=4.9.0
praw>=7.7.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
google-api-python-client>=2.100.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# LLM
anthropic>=0.18.0
openai>=1.0.0
tiktoken>=0.5.0

# Processing
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=23.1.0
tenacity>=8.2.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Web interface
fastapi>=0.104.0
uvicorn>=0.24.0

# Email
aiosmtplib>=3.0.0
